more code organization clean up
