+++ src/recipes/pages/Recipes.tsx
@@ -3,7 +3,8 @@ import { Route, Switch, useRouteMatch } from 'react-router-dom';
 import { RecipeBloc } from '../blocs/RecipeBloc';
 import { RecipeListBloc } from '../blocs/RecipeListBloc';
 import { Configuration, RecipesApi } from '../client';
-import { RecipeContextProvider, RecipeListContextProvider} from '../RecipesContext';
+import { RecipeContextProvider } from '../RecipeContext';
+import { RecipeListContextProvider } from '../RecipeListContext';
 import { RecipeListPage } from './RecipeListPage';
 import { RecipePage } from './RecipePage';
